inv adjustment trans working with detail

+++ guiclient/distributeInventory.cpp
@@ -26,6 +26,7 @@
 
 #define cIncludeLotSerial   0x01
 #define cNoIncludeLotSerial 0x02
+#define DEBUG false
 
 distributeInventory::distributeInventory(QWidget* parent, const char* name, bool modal, Qt::WindowFlags fl)
     : XDialog(parent, name, modal, fl)
@@ -84,7 +85,6 @@ distributeInventory::distributeInventory(QWidget* parent, const char* name, bool
 
   _itemlocdistid = -1;
   _transtype = "O";
-  
   _locationDefaultLit->hide();
   _locations->hide();
 }
@@ -134,10 +134,12 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
     itemloc.exec();
     while (itemloc.next())
     {
+      // Requires ot/serial
       if (itemloc.value("itemlocdist_reqlotserial").toBool())
       {
         int itemlocSeries = -1;
         XSqlQuery query;
+
         // Check to see if this is a lot controlled item and if we have
         // a predefined lot#/expdate to use. If so assign that information
         // with itemlocdist_qty and move on. otherwise do the normal dialog
@@ -172,50 +174,49 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
           }
         }
 
-        if(itemlocSeries == -1)
+        // Item did not have a preset lot number.
+        // Check to see if Lot/Serial distributions should be created using
+        // "from" side of transaction.  Transactions are related by itemlocdist_source_id.
+        // InterWarehouseTransfer uses this technique.
+        if(itemlocSeries == -1 && itemloc.value("itemlocdist_source_id").toInt() > -1)
         {
-          // Check to see if Lot/Serial distributions should be created using
-          // "from" side of transaction.  Transactions are related by itemlocdist_source_id.
-          // InterWarehouseTransfer uses this technique.
-          if (itemloc.value("itemlocdist_source_id").toInt() > -1)
+          XSqlQuery fromlots;
+          fromlots.exec("SELECT nextval('itemloc_series_seq') AS _itemloc_series;");
+          if(fromlots.first())
           {
-            XSqlQuery fromlots;
-            fromlots.exec("SELECT nextval('itemloc_series_seq') AS _itemloc_series;");
-            if(fromlots.first())
+            itemlocSeries = fromlots.value("_itemloc_series").toInt();
+            fromlots.prepare("SELECT  createlotserial(s.itemlocdist_itemsite_id, ls_number, "
+                             "        :itemlocdist_series, 'I', NULL, :itemlocdist_id, (d.itemlocdist_qty * -1.0), "
+                             "        itemloc_expiration, itemloc_warrpurc) "
+                             "FROM itemlocdist s JOIN itemlocdist o ON (o.itemlocdist_id=s.itemlocdist_source_id) "
+                             "                   JOIN itemlocdist d ON (d.itemlocdist_itemlocdist_id=o.itemlocdist_id) "
+                             "                   JOIN itemloc ON (itemloc_id=d.itemlocdist_source_id) "
+                             "                   JOIN ls ON (ls_id=itemloc_ls_id) "
+                             "WHERE (s.itemlocdist_id=:itemlocdist_id);"
+
+                             "UPDATE itemlocdist "
+                             "SET itemlocdist_source_type='O' "
+                             "WHERE (itemlocdist_series=:itemlocdist_series);"
+
+                             "DELETE FROM itemlocdist "
+                             "WHERE (itemlocdist_id=:itemlocdist_id);" );
+            fromlots.bindValue(":itemlocdist_series", itemlocSeries);
+            fromlots.bindValue(":itemlocdist_id", itemloc.value("itemlocdist_id"));
+            fromlots.exec();
+            if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Retrieving Lot/Serial Information"),
+                                          fromlots, __FILE__, __LINE__))
             {
-              itemlocSeries = fromlots.value("_itemloc_series").toInt();
-              fromlots.prepare("SELECT  createlotserial(s.itemlocdist_itemsite_id, ls_number, "
-                               "        :itemlocdist_series, 'I', NULL, :itemlocdist_id, (d.itemlocdist_qty * -1.0), "
-                               "        itemloc_expiration, itemloc_warrpurc) "
-                               "FROM itemlocdist s JOIN itemlocdist o ON (o.itemlocdist_id=s.itemlocdist_source_id) "
-                               "                   JOIN itemlocdist d ON (d.itemlocdist_itemlocdist_id=o.itemlocdist_id) "
-                               "                   JOIN itemloc ON (itemloc_id=d.itemlocdist_source_id) "
-                               "                   JOIN ls ON (ls_id=itemloc_ls_id) "
-                               "WHERE (s.itemlocdist_id=:itemlocdist_id);"
-
-                               "UPDATE itemlocdist "
-                               "SET itemlocdist_source_type='O' "
-                               "WHERE (itemlocdist_series=:itemlocdist_series);"
-
-                               "DELETE FROM itemlocdist "
-                               "WHERE (itemlocdist_id=:itemlocdist_id);" );
-              fromlots.bindValue(":itemlocdist_series", itemlocSeries);
-              fromlots.bindValue(":itemlocdist_id", itemloc.value("itemlocdist_id"));
-              fromlots.exec();
-              if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Retrieving Lot/Serial Information"),
-                                            fromlots, __FILE__, __LINE__))
-              {
-                return XDialog::Rejected;
-              }
+              return XDialog::Rejected;
             }
           }
         }
 
+        // Open Assign Lot Serial dialog, populating with auto ls info if required. 
         if(itemlocSeries == -1)
         { 
           ParameterList params;
           params.append("itemlocdist_id", itemloc.value("itemlocdist_id").toInt());
-
+          
           // Auto assign lot/serial if applicable
           if (itemloc.value("itemsite_lsseq_id").toInt() != -1 &&
               !itemloc.value("itemsite_perishable").toBool() &&
@@ -232,7 +233,6 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
               return XDialog::Rejected;
             }
           }
-
           assignLotSerial newdlg(pParent, "", true);
           newdlg.set(params);
           itemlocSeries = newdlg.exec();
@@ -240,6 +240,7 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
             return XDialog::Rejected;
         }
         
+        // Distribute location with Distribute Inventory dialog
         if (itemloc.value("itemsite_loccntrl").toBool())
         {
           query.prepare( "SELECT itemlocdist_id " 
@@ -281,7 +282,7 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
         else
         {
           query.prepare( "UPDATE itemlocdist "
-                         "SET itemlocdist_source_type='L', itemlocdist_source_id=-1 "
+                         "SET itemlocdist_source_type='L', itemlocdist_source_id = -1 "
                          "WHERE (itemlocdist_series=:itemlocdist_series); ");
           query.bindValue(":itemlocdist_series", itemlocSeries);
           query.exec();
@@ -289,15 +290,36 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
           // Append id to list and process at the end
           ildsList.append(itemlocSeries);
         }
+
+        // Set itemlocdist_child_series of parent itemlocdist record. Can this be removed since it's below?
+        query.prepare("UPDATE itemlocdist SET itemlocdist_child_series = :itemlocSeries "
+                      "WHERE itemlocdist_series = :pItemlocSeries "
+                      " AND :itemlocSeries IS NOT NULL "
+                      "RETURNING itemlocdist_id;");
+        query.bindValue(":itemlocSeries", itemlocSeries);
+        query.bindValue(":pItemlocSeries",  pItemlocSeries);
+        query.exec();
+        if (query.first())
+        {
+          if (query.count() != 1)
+          {
+            ErrorReporter::error(QtCriticalMsg, 0, tr("Updating Itemlocdist Parent Record Should Return One Row"),
+                           query, __FILE__, __LINE__);
+            return XDialog::Rejected;
+          }
+        }  
+        else if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Updating itemlocdist Lot/Serial Information"),
+                                      query, __FILE__, __LINE__))
+          return XDialog::Rejected;
       }
-      else
+      else // NOT a lot/serial controlled item
       {
         ParameterList params;
         params.append("itemlocdist_id", itemloc.value("itemlocdist_id").toInt());
         params.append("trans_type", itemloc.value("trans_type").toString());
 
         if (itemloc.value("itemlocdist_distlotserial").toBool())
-          params.append("includeLotSerialDetail");
+          params.append("includeLotSerialDetail", true); // true arg required here?
 
         distributeInventory newdlg(pParent, "", true);
         newdlg.set(params);
@@ -324,50 +346,58 @@ int distributeInventory::SeriesAdjust(int pItemlocSeries, QWidget *pParent, cons
         }
       }
     }
-
-    XSqlQuery post;
     
-    // Process Lot/Serial distributions
-    for (int i = 0; i < ildsList.size(); ++i) {
-      post.prepare( "SELECT distributeItemlocSeries(:itemlocdist_series) AS result;");
+    XSqlQuery post;
+    for (int i = 0; i < ildsList.size(); ++i) { // Process Lot/Serial distributions
+      if (DEBUG)
+        qDebug() << "DistributeInventory::seriesAdjust ildsList.at(i): " << ildsList.at(i);
+
+      post.prepare( "UPDATE itemlocdist SET itemlocdist_child_series = :itemlocdist_series "
+                    "WHERE itemlocdist_series = :pItemlocSeries "
+                    "RETURNING itemlocdist_id;");
+      post.bindValue(":pItemlocSeries", pItemlocSeries);
       post.bindValue(":itemlocdist_series", ildsList.at(i));
       post.exec();
-      if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Retrieving Lot/Serial Information"),
-                                    post, __FILE__, __LINE__))
+      if (post.first())
       {
+        if (post.count() != 1)
+        {
+          ErrorReporter::error(QtCriticalMsg, 0, tr("Updating Itemlocdist Parent Record Should Return One Row"),
+                         post, __FILE__, __LINE__);
+          return XDialog::Rejected;
+        }
+      }  
+      else if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Updating itemlocdist Lot/Serial Information"),
+                                    post, __FILE__, __LINE__))
         return XDialog::Rejected;
-      }
     }
     
-    // Process location distributions
-    for (int i = 0; i < ildList.size(); ++i) {
-      post.prepare("SELECT distributeToLocations(:itemlocdist_id) AS result;");
+    for (int i = 0; i < ildList.size(); ++i) { // Process location distributions
+      if (DEBUG)
+        qDebug() << "DistributeInventory::seriesAdjust ildsList.at(i): " << ildsList.at(i);
+
+      post.prepare( "UPDATE itemlocdist SET itemlocdist_child_series = ( "
+                    " SELECT itemlocdist_series FROM itemlocdist "
+                    " WHERE itemlocdist_id = :itemlocdist_id) "
+                    "WHERE itemlocdist_series = :pItemlocSeries "
+                    "RETURNING itemlocdist_id;");
       post.bindValue(":itemlocdist_id", ildList.at(i));
+      post.bindValue(":pItemlocSeries", pItemlocSeries);
       post.exec();
-      if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Retrieving Lot/Serial Information"),
-                                    post, __FILE__, __LINE__))
+      if (post.first())
       {
-        return XDialog::Rejected;
+        if (post.count() <= 0)
+        {
+          ErrorReporter::error(QtCriticalMsg, 0, tr("Updating Itemlocdist Parent Record Should Return >= 1"),
+                         post, __FILE__, __LINE__);
+          return XDialog::Rejected;
+        }
       }
-    }
-    
-    //Post inventory history for any remaining non-distributed transactions and trial balance
-    post.prepare("SELECT postItemlocseries(:itemlocseries) AS result;");
-    post.bindValue(":itemlocseries",  pItemlocSeries);
-    post.exec();
-    if (post.first())
-    {
-      if (!post.value("result").toBool())
-            QMessageBox::warning( 0, tr("Inventory Distribution"), 
-        tr("There was an error posting the transaction.  Contact your administrator") );
-    }
-    else if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Retrieving Lot/Serial Information"),
-                                  post, __FILE__, __LINE__))
-    {
-      return XDialog::Rejected;
+      else if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Updating itemlocdist Lot/Serial Information"),
+                                    post, __FILE__, __LINE__))
+        return XDialog::Rejected;
     }
   }
-  
   return XDialog::Accepted;
 }
 
@@ -428,7 +458,7 @@ void distributeInventory::populate()
     else
       _bcQty->clear();
   }
-  else if (ErrorReporter::error(QtCriticalMsg, this, tr("Error Retrieving Lot/Serial Information"),
+  else if (ErrorReporter::error(QtCriticalMsg, 0, tr("Error Retrieving Lot/Serial Information"),
                                 distributepopulate, __FILE__, __LINE__))
   {
     return;
